refactor(runtime): separate status projection from session ownership
